refactor advisory workspace version lookups
